+++ NEWS.rst
@@ -6,7 +6,8 @@ of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added cxx_gxx103_armv7 mkspec for GCC 10.3 ARM cross compiler.
+* Minor: Added cxx_gxx103_aarch64 mkspec for GCC 10.3 ARM64 cross compiler.
 
 4.40.0
 ------